refactor(login): use provider to access static instances
In preparation for Fake testing, uses providers to inject into the
`LoginService`.
