--- lib/services/implementations/login_service_firebase.dart
@@ -1,29 +1,0 @@
-import "package:firebase_auth/firebase_auth.dart";
-import "package:google_sign_in/google_sign_in.dart";
-import "package:proxima/services/login_service.dart";
-
-class LoginServiceFirebase extends LoginService {
-  @override
-  Future<void> signIn() async {
-    // Trigger the authentication flow
-    final GoogleSignInAccount? googleUser = await GoogleSignIn().signIn();
-
-    // Obtain the auth details from the request
-    final GoogleSignInAuthentication? googleAuth =
-        await googleUser?.authentication;
-
-    // Create a new credential
-    final credential = GoogleAuthProvider.credential(
-      accessToken: googleAuth?.accessToken,
-      idToken: googleAuth?.idToken,
-    );
-
-    // Once signed in, return the UserCredential
-    await FirebaseAuth.instance.signInWithCredential(credential);
-  }
-
-  @override
-  Future<void> signOut() async {
-    await FirebaseAuth.instance.signOut();
-  }
-}
