+++ lib/utils/ui/circular_value.dart
@@ -1,7 +1,7 @@
 import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 
-/// Widget used to display a [CircularProgressIndicator] while waiting for an [AsyncValue] to complete; and another widget once the data resolves.
+/// Utilitiy widget used to display a [CircularProgressIndicator] while waiting for an [AsyncValue] to complete; and another widget once the data resolves.
 class CircularValue<T> extends StatelessWidget {
   final AsyncValue<T> value;
   final Widget Function(BuildContext context, T data) builder;
