+++ lib/viewmodels/login_viewmodel.dart
@@ -1,9 +1,6 @@
 import "package:firebase_auth/firebase_auth.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
-
 import "package:proxima/models/login_user.dart";
-import "package:proxima/models/ui/profile_data.dart";
-import "package:proxima/services/implementations/login_service_firebase.dart";
 import "package:proxima/services/login_service.dart";
 
 /// Firebase authentication change provider
@@ -17,29 +14,10 @@ final userProvider = StreamProvider<LoginUser?>((ref) {
   });
 });
 
-/// User profile view model
-class ProfileViewModel extends AsyncNotifier<ProfileData> {
-  ProfileViewModel();
-
-  @override
-  Future<ProfileData> build() async {
-    final user = ref.watch(userProvider).valueOrNull;
-    if (user == null) {
-      return Future.error(
-        "User must be logged in before displaying the home page.",
-      );
-    }
-
-    return ProfileData(user: user);
-  }
-}
-
-/// Login Service provider, dependency injection
-final loginServiceProvider = Provider<LoginService>((_) {
-  return LoginServiceFirebase();
+/// Login Service provider; dependency injection used for testing purposes
+final loginServiceProvider = Provider<LoginService>((ref) {
+  return LoginService(
+    firebaseAuth: ref.watch(firebaseAuthProvider),
+    googleSignIn: ref.watch(googleSignInProvider),
+  );
 });
-
-/// Profile view model of the currently logged in user
-final profileProvider = AsyncNotifierProvider<ProfileViewModel, ProfileData>(
-  () => ProfileViewModel(),
-);
