+++ lib/views/pages/home_page.dart
@@ -2,6 +2,7 @@ import "package:flutter/material.dart";
 import "package:hooks_riverpod/hooks_riverpod.dart";
 import "package:proxima/utils/ui/circular_value.dart";
 import "package:proxima/viewmodels/login_viewmodel.dart";
+import "package:proxima/viewmodels/profile_view_model.dart";
 import "package:proxima/views/navigation/routes.dart";
 
 class HomePage extends HookConsumerWidget {
